make it work with route data
