+++ examples/hackernews-ts/src/routes/users/[id].tsx
@@ -9,8 +9,18 @@ export interface IUser {
   about: string;
 }
 
+import { createResource } from "solid-js";
+import { RouteDataFunc } from "solid-app-router";
+import fetchAPI from "../../lib/api";
+import server from "solid-start/server";
+
+export const routeData: RouteDataFunc = props => {
+  const [user] = createResource(() => `user/${props.params.id}`, server(fetchAPI));
+  return user;
+};
+
 const User: Component = () => {
-  const user = useData<() => IUser>()
+  const user = useData<() => IUser>();
   return (
     <div class="user-view">
       <Show when={user()}>
